refactor(sdk): extract Guardian inspector to _client/_guardian.py
PR 2/N of the SDK refactor (see #702 for the WebSocket pilot and the
mixin template). Moves the ADR-016 Phase 3 Guardian cooperation hook
out of cullis_sdk/client.py into a dedicated mixin so client.py
keeps shrinking.

Guardian is the second-most-isolated SDK domain after WebSocket: four
methods plus the section header, no overlap with auth, sessions,
messaging or AI gateway. Only call sites are inside send_oneshot and
the two reply paths — those continue to resolve through the mixin.

What moves:

- cullis_sdk/_client/_guardian.py — new module exposing
    * _GuardianMixin with:
        _guardian_inspect
        _guardian_inspect_send
        _guardian_inspect_deliver
        _apply_guardian_deliver
  Imports of InspectionDecision, _build_body, _enabled, _no_op_decision
  and _parse_response from cullis_sdk.guardian.client are scoped to
  this module.
- cullis_sdk/client.py — CullisClient now inherits _GuardianMixin and
  _WebSocketMixin (alphabetical order). The four methods, their
  section comment, and the now-unused guardian-related imports are
  removed from client.py. The GuardianBlocked import is dropped too —
  it was referenced only in docstrings/comments, never raised inside
  client.py (the exception is raised inside cullis_sdk.guardian.client
  itself and propagates up).

Public surface is unchanged. Guardian methods are private, so no
__init__.py update is required.

No logic change — pure code movement.
